renamed gamepage-file, added basic consts
