+++ packages/client/src/pages/Game/index.ts
@@ -1,1 +1,1 @@
-export { default as Game } from './Game'
+export { default as Game } from './GamePage'
